add --ignore-missing to ignore missing input files
When feeding filenames from 'find' to tar, it is possible for files to
disappear before being processed. This forces tar to signal an error
unless --ignore-failed-read has been specified. Using this switch
however covers a wide variety of error conditions. To simplify the
archival of big filesystems with find and tar, this change adds the
command line switch "--ignore-missing" to ignore files that disappeared
in the time between find identifying them and tar processing them.